Some more refactoring.
Moved ViewPet to be a dynamic route based off of animal id.
