+++ src/app/ViewPet/[id]/page.js
@@ -1,9 +1,9 @@
 import React from "react";
-import Navbar from "../components/Navbar";
+import Navbar from "@components/Navbar";
 import Image from "next/image";
 import Link from "next/link";
 
-const Page = () => {
+const Page = ({params}) => {
   return (
     <div>
       <Navbar />
@@ -13,7 +13,7 @@ const Page = () => {
           <div className="flex flex-col items-center justify-center w-1/2 p-8 mt-10">
             {/* Moves the image inside the left side div */}
             <div className="w-80 h-80 relative">
-              <Link href="/BrowsePets?category=Dogs">
+              <Link href="/BrowsePets?category=Animals">
                 <button className="text-sm rounded p-2 bg-white text-black mb-4 hover:bg-orange-500">
                   &lt; Back to Listings
                 </button>
@@ -36,8 +36,9 @@ const Page = () => {
           <div className="w-1/2 p-8 mt-20">
             <h1 className="text-4xl font-bold mb-8">Animal Name</h1>
             <h2 className="text-2xl font-bold mb-4">
-              Animal ID:{" "}
+              Animal ID:{params.id}
               <span className="text-orange-500 font-bold">XXXXXXXXXXX</span>
+              {/*  Not sure why we need the <span> above */}
             </h2>
             <p className="text-lg">Age: XXX</p>
             <p className="text-lg mb-2">Category: XXX</p>
